Plugins (nastyfft): Fix a fail in gnome-settings-daemon (LP: #287448)
The static VisPluginInfo in nastyfft was declared as const, but libvisual
write to its refcount when it visual_object_unref()s it (nasty!).  This is
broken at various levels: libvisual shouldn't be writing to this static
plugin which is allocated by a dlopen()ed plugin, and it should also honor
the "const VisPluginInfo" API...  Anyway, nastyfft was the only occurrence
and this fixes the crash for now; LP: #287448.