libvisual/lv_audio.c: Fix input_interleaved_stereo logic and crashes
Previously, twice as many bytes as available were read and written
(heap-buffer-overflow) and the code was not robust towards and
odd number of samples. Thanks to AddressSanitizer!

The two calls to visual_buffer_set_destroyer were dropped because
visual_buffer_new_allocate in macro STEREO_INTERLEAVED is already
setting visual_buffer_destroyer_free as a destroyer.